Add crop node
In this change crop node is implemented.
Crop node expects 4 inputs:
- First node input
- Second node input
- x offset
- y offset
The purpose of the node is to crop portion of the first node that
dimension-wise corresponds to the second node at the given offset.
Currently, offsets must be given manually; in future they will be
calculated based on affine transform between two inputs along the
path that connects them.